LayoutTests/http/tests/websocket/tests/hybi/non-document-mixed-content-blocked-https-with-embedded-http-with-embedded-https-UpgradeMixedContent.https-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/websocket/tests/hybi/non-document-mixed-content-blocked-https-with-embedded-http-with-embedded-https-UpgradeMixedContent.https.html requested insecure content from http://localhost:8000/websocket/tests/hybi/non-document-mixed-content-blocked-http-with-embedded-https.html. This content was blocked and must be served over HTTPS.

PASS successfullyParsed is true

TEST COMPLETE
PASS Timeout while waiting for insecure iframe

Test topFrame https with embedded http with embedded https blocks worker trying to connect with insecure webSocket.
